requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pymongo>=4.6.0
motor>=3.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
